requirements.txt
Django>=1.10.4
django-webpack-loader==0.4.1
djangorestframework==3.5.3
django-debug-toolbar
psycopg2
